feat: move service into mods
