+++ service/src/internal/job.rs
@@ -1,0 +1,77 @@
+use base64::Engine;
+use celestia_types::{blob::Commitment, block::Height as BlockHeight, nmt::Namespace};
+use eq_common::{InclusionServiceError, KeccakInclusionToDataRootProofInput};
+use serde::{Deserialize, Serialize};
+use sp1_sdk::SP1ProofWithPublicValues;
+
+use crate::SuccNetJobId;
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct Job {
+    pub height: BlockHeight,
+    pub namespace: Namespace,
+    pub commitment: Commitment,
+}
+
+impl Job {
+    pub fn new(height: BlockHeight, namespace: Namespace, commitment: Commitment) -> Self {
+        Job {
+            height,
+            namespace,
+            commitment,
+        }
+    }
+}
+
+impl std::fmt::Debug for Job {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let namespace_string;
+        if let Some(namespace) = &self.namespace.id_v0() {
+            namespace_string = base64::engine::general_purpose::STANDARD.encode(namespace);
+        } else {
+            namespace_string = "Invalid v0 ID".to_string()
+        }
+        let commitment_string =
+            base64::engine::general_purpose::STANDARD.encode(&self.commitment.hash());
+        f.debug_struct("Job")
+            .field("height", &self.height.value())
+            .field("namespace", &namespace_string)
+            .field("commitment", &commitment_string)
+            .finish()
+    }
+}
+
+/// Used as a [Job] state machine for the eq-service.
+///
+/// Should map 1to1 with [ResponseStatus] for consistency in internal state
+/// and what is reported by the RPC.
+#[derive(Serialize, Deserialize)]
+pub enum JobStatus {
+    /// DA inclusion proof data is being collected
+    DataAvailabilityPending,
+    /// DA inclusion is processed and ready to send to the ZK prover
+    DataAvailable(KeccakInclusionToDataRootProofInput),
+    /// A ZK prover job had been requested, awaiting response
+    ZkProofPending(SuccNetJobId),
+    /// A ZK proof is ready, and the [Job] is complete
+    // For now we'll use the SP1ProofWithPublicValues as the proof
+    // Ideally we only want the public values + whatever is needed to verify the proof
+    // They don't seem to provide a type for that.
+    ZkProofFinished(SP1ProofWithPublicValues),
+    /// A wrapper for any [InclusionServiceError], with:
+    /// - Option = None                        --> Permanent failure
+    /// - Option = Some(\<retry-able status\>) --> Retry is possible, with a JobStatus state to retry with
+    Failed(InclusionServiceError, Option<Box<JobStatus>>),
+}
+
+impl std::fmt::Debug for JobStatus {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            JobStatus::DataAvailabilityPending => write!(f, "DataAvailabilityPending"),
+            JobStatus::DataAvailable(_) => write!(f, "DataAvailable"),
+            JobStatus::ZkProofPending(_) => write!(f, "ZkProofPending"),
+            JobStatus::ZkProofFinished(_) => write!(f, "ZkProofFinished"),
+            JobStatus::Failed(_, _) => write!(f, "Failed"),
+        }
+    }
+}
